requirements.txt
asgiref==3.3.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
djangorestframework==3.15.1
idna==2.10
importlib-resources==5.4.0
lxml==4.6.3
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==4.1.1
urllib3==1.26.6
zipp==3.6.0
